full year (#11)
* refactor: Use constant for uncleared status and improve help toggle logic

* Refactor message handling into separate functions

The motivation for the change is to improve code readability and maintainability by refactoring the message handling logic into separate functions. This change makes the codebase easier to understand and modify, as each function now handles a specific type of message.

Previously, the message handling logic was embedded directly within the `Update` function, making it lengthy and harder to follow. By moving the logic into dedicated functions like `handleWindowSize`, `handleGetUser`, `handleGetTransactions`, etc., the `Update` function is now cleaner and more focused on message dispatching. This refactoring also promotes code reuse and separation of concerns, as each function is responsible for handling a specific message type.

* feat: Introduce Period struct for date range management and formatting

* refactor: Replace string period with Period struct and update transaction date handling

* fix: Clamp period end date to the last day of the current month

* refactor: Extract period calculation to setPeriod method in Period struct

* feat: Add default case for setting period in setPeriod function

* feat: Add rangeType to model for switching between month and year

* Refactor period type handling

Motivation for the change:
The motivation for this change is to improve code readability and maintainability by using more descriptive variable names and constants for period types. This helps in understanding the code better and reduces the risk of errors when handling different period types.

How it differs from the previous implementation:
- Rename `switchRange` to `switchPeriod` for clarity.
- Introduce constants `monthlyPeriodType` and `annualPeriodType` to replace hardcoded strings "month" and "year".
- Update variable `rangeType` to `periodType` to better reflect its purpose.
- Modify the `setPeriod` function to use the new constants and provide a default case for period type.

* feat: Automatically move cursor down to the next transaction item